Attach XFF header to request to downstream
Previously, djproxy was erroneously calculating its own IP address and attaching an XFF header to the response it sent back to the client. That wasn't even remotely the right behavior.

This changes things so that djproxy updates (or adds) an xff header with the client's IP address to the request it is sending to the proxied endpoint.