Dislabe DATA_UPLOAD_MAX_MEMORY_SIZE checks
There is a bug in Django 1.10 which causes it to blow up if the CONTENT_LENGTH header is set to an empty string. This has been corrected in master and will likely ship in django 1.10.

For some configurations, it may be necessary to disable upload max memory size checking (which was introduced in 1.10).